Lesson 109

+++ spring7restmvc/src/main/java/guru/springframework/spring7restmvc/service/impl/BeerServiceJPA.java
@@ -61,8 +61,12 @@ public class BeerServiceJPA implements BeerService {
 	}
 
 	@Override
-	public void deleteById(UUID beerId) {
+	public boolean deleteById(UUID beerId) {
+		if(!beerRepository.existsById(beerId)) {
+			return false;
+		}
 		beerRepository.deleteById(beerId);
+		return true;
 	}
 
 	@Override
